requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
mysqlclient==2.2.7
stripe==10.12.0
celery==5.6.2
redis==7.3.0
django-environ==0.13.0
django-cors-headers==4.9.0
django-redis==6.0.0
django-celery-beat==2.9.0
gunicorn==25.1.0
whitenoise==6.12.0
